When the title of a post is edited, display an alert message to the user and have them choose whether to update the slug or keep the existing one. Update the post accordingly.

+++ BlogTemplate/Models/BlogDataStore.cs
@@ -293,7 +293,10 @@ namespace BlogTemplate.Models
         public void UpdatePost(Post newPost, Post oldPost)
         {
             SavePost(newPost);
-            _fileSystem.DeleteFile($"{StorageFolder}\\{oldPost.Slug}.xml");
+            if (newPost.Slug != oldPost.Slug)
+            {
+                _fileSystem.DeleteFile($"{StorageFolder}\\{oldPost.Slug}.xml");
+            }
         }
 
         public bool CheckSlugExists(string slug)
